+++ src/components/generation/carousal.tsx
@@ -48,13 +48,38 @@ const CurvedEmblaCarousel = ({
     // Update the ref to track changes
     prevLengthRef.current = modelRequest.length;
   }, [modelRequest.length, api]);
+  const ref = useRef<HTMLDivElement>(null);
+  useEffect(() => {
+    const modelContainer = ref.current;
+    if (!modelContainer) return;
+    const stopPropagation = (e: Event) => {
+      e.stopPropagation();
+    };
+    modelContainer.addEventListener("mousedown", stopPropagation);
+    modelContainer.addEventListener("touchcancel", stopPropagation);
+    modelContainer.addEventListener("touchend", stopPropagation);
+    modelContainer.addEventListener("touchmove", stopPropagation);
+    modelContainer.addEventListener("wheel", stopPropagation);
+    modelContainer.addEventListener("touchstart", stopPropagation);
+    return () => {
+      if (!modelContainer) return;
+      modelContainer.removeEventListener("mousedown", stopPropagation);
+      modelContainer.removeEventListener("touchcancel", stopPropagation);
+      modelContainer.removeEventListener("touchend", stopPropagation);
+      modelContainer.removeEventListener("touchmove", stopPropagation);
+      modelContainer.removeEventListener("wheel", stopPropagation);
+      modelContainer.removeEventListener("touchstart", stopPropagation);
+    };
+  }, [ref]);
+
   return (
     <div
-      className="max-w-[264px] max-h-[370px] h-full w-full"
+      className="max-w-[264px] max-h-[370px] h-full w-full "
       ref={carouselRef}
     >
       {modelRequest.length ? (
         <Carousel
+          ref={ref}
           opts={{
             align: "center",
             startIndex:
@@ -69,24 +94,6 @@ const CurvedEmblaCarousel = ({
           plugins={[WheelGesturesPlugin()]}
           className="overflow-visible  pointer-events-none relative max-w-[264px] max-h-[370px] h-full w-full"
           setApi={setApi}
-          onMouseDown={(e) => {
-            e.stopPropagation();
-          }}
-          onTouchCancel={(e) => {
-            e.stopPropagation();
-          }}
-          onTouchEnd={(e) => {
-            e.stopPropagation();
-          }}
-          onTouchMove={(e) => {
-            e.stopPropagation();
-          }}
-          onTouchStart={(e) => {
-            e.stopPropagation();
-          }}
-          onWheel={(e) => {
-            e.stopPropagation();
-          }}
         >
           <CarouselContent
             className="ml-0 relative h-full pointer-events-none w-full"
@@ -120,8 +127,12 @@ const CurvedEmblaCarousel = ({
                     style={{
                       transform: `translate(${transformX}, ${transformY}) rotate(${rotate}) scale(${scale})`,
                     }}
+                    onClick={() => {
+                      if (isCurrent) return;
+                      api?.scrollTo(index);
+                    }}
                     className={cn(
-                      "relative pl-0  border-buu  pointer-events-none select-none max-w-[264px]  max-h-[370px] h-full rounded-lg shadow-lg",
+                      "relative pl-0  border-buu   pointer-events-none select-none max-w-[264px]  max-h-[370px] h-full rounded-lg shadow-lg",
                       "transition-all duration-500 ease-out",
                     )}
                   >