+++ src/components/generation/tool-bar-tool-tips.tsx
@@ -14,16 +14,25 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { motion } from "framer-motion";
 import toast from "react-hot-toast";
 import { ToolTips, TToolTipEvents } from "./handle-tool-calls";
-import ToolTipModify from "./tool-tip-modify";
+// import ToolTipModify from "./tool-tip-modify";
+import ToolTipDownload from "./tool-tip-download";
 
 type TToolBarToolTips = {
   subThreadId: string;
   imageUrl: string | null;
+  modelUrl?: string | null;
+};
+export const buttonVariants = {
+  initial: { y: 0, scale: 1 },
+  // hover: { y: -2, scale: 1.05 },
+  tap: { y: 0, scale: 0.95 },
+  drag: { scale: 1.1, boxShadow: "0px 5px 10px rgba(0, 0, 0, 0.2)" },
 };
 
 export default function ToolBarToolTips({
   subThreadId,
-  imageUrl,
+  // imageUrl,
+  modelUrl,
 }: TToolBarToolTips) {
   const dispatch = useAppDispatch();
   const { getAccessToken } = useAuthentication();
@@ -45,10 +54,6 @@ export default function ToolBarToolTips({
     },
   });
 
-  // function handleModifyDrag() {
-  //   // Set the dragged image URL in Redux store
-  //   dispatch(setDraggedImage(imageUrl ?? undefined));
-  // }
 
   function handleEvent(events: TToolTipEvents) {
     const accessToken = getAccessToken(address ?? "");
@@ -64,10 +69,9 @@ export default function ToolBarToolTips({
         });
         break;
       }
-      case "MODIFY": {
-        toast.loading("Generating model...");
-        break;
-      }
+      // case "MODIFY": {
+      //   break;
+      // }
       default: {
         console.log("NOT FOUND");
       }
@@ -75,25 +79,33 @@ export default function ToolBarToolTips({
   }
 
   // Define button hover animation variants
-  const buttonVariants = {
-    initial: { y: 0, scale: 1 },
-    hover: { y: -2, scale: 1.05 },
-    tap: { y: 0, scale: 0.95 },
-    drag: { scale: 1.1, boxShadow: "0px 5px 10px rgba(0, 0, 0, 0.2)" },
-  };
 
   return (
     <TooltipProvider>
       {ToolTips.map((item, index) => {
-        if (item.type === "MODIFY" && imageUrl) {
+        // if (item.type === "MODIFY" && imageUrl) {
+        //   return (
+        //     <ToolTipModify
+        //       key={`tool-tip-contents-${item.content.trim()}-${subThreadId}-${index}`}
+        //       index={index}
+        //       length={ToolTips.length}
+        //       subThreadId={subThreadId}
+        //       toolTipData={item}
+        //       imageUrl={imageUrl}
+        //     />
+        //   );
+        // }
+
+        if (item.type === "DOWNLOAD") {
+          // key={`tool-tip-contents-${item.content.trim()}-${index}`}
           return (
-            <ToolTipModify
-              key={`tool-tip-contents-${item.content.trim()}-${subThreadId}-${index}`}
+            <ToolTipDownload
+              key={`tool-tip-contents-${item.content.trim()}-${index}`}
+              modelUrl={modelUrl}
               index={index}
               length={ToolTips.length}
               subThreadId={subThreadId}
               toolTipData={item}
-              imageUrl={imageUrl}
             />
           );
         }
@@ -108,6 +120,7 @@ export default function ToolBarToolTips({
                 whileTap="tap"
                 variants={buttonVariants}
                 transition={{ type: "spring", stiffness: 400, damping: 17 }}
+                // group bg-buu-button  hover:bg-white hover:shadow-none  group shadow-buu-button min-w-[30px]  rounded-md flex items-center justify-center p-1.5
                 className="group bg-buu-button pointer-events-auto hover:bg-white hover:shadow-none group shadow-buu-button min-w-[30px] rounded-md flex items-center justify-center p-1.5"
               >
                 <motion.div